# Conflicts:
#	core/src/main/java/io/questdb/cutlass/http/HttpConnectionContext.java
#	core/src/test/java/io/questdb/cutlass/http/IODispatcherTest.java

+++ core/src/main/java/io/questdb/griffin/engine/groupby/vect/AvgDoubleVectorAggregateFunction.java
@@ -74,7 +74,7 @@ public class AvgDoubleVectorAggregateFunction extends DoubleFunction implements 
     @Override
     public void initRosti(long pRosti) {
         Unsafe.getUnsafe().putDouble(Rosti.getInitialValueSlot(pRosti, this.valueOffset), 0);
-        Unsafe.getUnsafe().putDouble(Rosti.getInitialValueSlot(pRosti, this.valueOffset + 1), 0);
+        Unsafe.getUnsafe().putLong(Rosti.getInitialValueSlot(pRosti, this.valueOffset + 1), 0);
     }
 
     @Override
